Rearranged Options and Arguments based on what would most commonly be used
The MyBB version to be installed is now an option because most people will want to install the latest version. The name of the folder in which to install is now the argument since most people will want to install into a specific directory. Just as before though, if no directory name is given, it will install into the current directory.